Move some code from MainActivity to WalletConnect
